+++ routes/web.php
@@ -1,5 +1,4 @@
 <?php
 
-Route::get('/{any}', function () {
-    return view('laravel-enso/core::index');
-})->where('any', '.*');
+Route::view('/{any}', 'laravel-enso/core::index')
+    ->where('any', '.*');